Revert "Use key prefix for http cache for BuildCacheNG"
This reverts commit d909e9dd5d62aa5cc5663d584d80a77a3e4e316d.

Revert "Add integration test for BuildCacheNG http key prefix"

This reverts commit 0ad88c599e35de65c85cb0e1b5eda0d063a93b37.
