+++ subprojects/build-cache-http/src/main/java/org/gradle/caching/http/internal/HttpBuildCacheService.java
@@ -46,8 +46,11 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Set;
 
-public abstract class AbstractHttpBuildCacheService implements BuildCacheService {
-    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractHttpBuildCacheService.class);
+/**
+ * Build cache implementation that delegates to a service accessible via HTTP.
+ */
+public class HttpBuildCacheService implements BuildCacheService {
+    private static final Logger LOGGER = LoggerFactory.getLogger(HttpBuildCacheService.class);
     static final String BUILD_CACHE_CONTENT_TYPE = "application/vnd.gradle.build-cache-artifact.v1";
 
     private static final Set<Integer> FATAL_HTTP_ERROR_CODES = ImmutableSet.of(
@@ -66,7 +69,7 @@ public abstract class AbstractHttpBuildCacheService implements BuildCacheService
     private final HttpBuildCacheRequestCustomizer requestCustomizer;
     private final boolean useExpectContinue;
 
-    public AbstractHttpBuildCacheService(HttpClientHelper httpClientHelper, URI url, HttpBuildCacheRequestCustomizer requestCustomizer, boolean useExpectContinue) {
+    public HttpBuildCacheService(HttpClientHelper httpClientHelper, URI url, HttpBuildCacheRequestCustomizer requestCustomizer, boolean useExpectContinue) {
         this.requestCustomizer = requestCustomizer;
         this.useExpectContinue = useExpectContinue;
         this.root = withTrailingSlash(url);
@@ -75,7 +78,7 @@ public abstract class AbstractHttpBuildCacheService implements BuildCacheService
 
     @Override
     public boolean contains(BuildCacheKey key) {
-        final URI uri = resolveUriForKey(root, key);
+        final URI uri = root.resolve("./" + key.getHashCode());
         HttpHead httpHead = new HttpHead(uri);
         requestCustomizer.customize(httpHead);
 
@@ -100,7 +103,7 @@ public abstract class AbstractHttpBuildCacheService implements BuildCacheService
 
     @Override
     public boolean load(BuildCacheKey key, BuildCacheEntryReader reader) throws BuildCacheException {
-        final URI uri = resolveUriForKey(root, key);
+        final URI uri = root.resolve("./" + key.getHashCode());
         HttpGet httpGet = new HttpGet(uri);
         httpGet.addHeader(HttpHeaders.ACCEPT, BUILD_CACHE_CONTENT_TYPE + ", */*");
         requestCustomizer.customize(httpGet);
@@ -127,7 +130,7 @@ public abstract class AbstractHttpBuildCacheService implements BuildCacheService
 
     @Override
     public void store(BuildCacheKey key, final BuildCacheEntryWriter output) throws BuildCacheException {
-        final URI uri = resolveUriForKey(root, key);
+        final URI uri = root.resolve(key.getHashCode());
         HttpPut httpPut = new HttpPut(uri);
         if (useExpectContinue) {
             httpPut.setHeader(HTTP.EXPECT_DIRECTIVE, HTTP.EXPECT_CONTINUE);
@@ -177,8 +180,6 @@ public abstract class AbstractHttpBuildCacheService implements BuildCacheService
             throw wrap(e);
         }
     }
-
-    protected abstract URI resolveUriForKey(URI root, BuildCacheKey key);
 
     private static BuildCacheException wrap(Throwable e) {
         if (e instanceof Error) {
