--- subprojects/build-cache-http/src/main/java/org/gradle/caching/http/internal/NextGenHttpBuildCacheService.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.caching.http.internal;
-
-import org.gradle.caching.BuildCacheKey;
-import org.gradle.internal.resource.transport.http.HttpClientHelper;
-
-import java.net.URI;
-
-/**
- * Build cache implementation that delegates to a service accessible via HTTP for NextGen build cache.
- */
-public class NextGenHttpBuildCacheService extends AbstractHttpBuildCacheService {
-
-    public NextGenHttpBuildCacheService(HttpClientHelper httpClientHelper, URI url, HttpBuildCacheRequestCustomizer requestCustomizer, boolean useExpectContinue) {
-        super(httpClientHelper, url, requestCustomizer, useExpectContinue);
-    }
-
-    @Override
-    protected URI resolveUriForKey(URI root, BuildCacheKey key) {
-        return root.resolve("./nextGen-" + key.getHashCode());
-    }
-}
